requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
python-multipart>=0.0.6
sqlalchemy>=2.0.20
psycopg2-binary>=2.9.7
pydantic>=2.3.0
pydantic-settings>=2.0.3

# ML & OCR
easyocr>=1.7.1
pytesseract>=0.3.10
spacy>=3.6.1
transformers>=4.33.1
torch>=2.0.1
opencv-python-headless>=4.8.0

# Language detection
langdetect>=1.0.9

# Data synthesis
Faker>=19.6.0
reportlab>=4.0.5

# Utilities
python-dotenv>=1.0.0
pytest>=7.4.2
requests>=2.31.0

PyMuPDF>=1.23.0
